requirements.txt
blinker==1.7.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cycler==0.12.1
einops==0.7.0
filelock==3.9.0
Flask==3.0.2
fonttools==4.45.1
fsspec==2023.12.1
huggingface-hub==0.19.4
idna==3.4
imageio==2.33.0
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.5.0
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
opencv-python==4.7.0.72
packaging==23.2
Pillow==9.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.28.1
safetensors==0.4.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
setuptools-scm==8.0.4
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.12
tomli==2.0.1
torchsummary==1.5.1
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==3.0.1
zipp==3.17.0
